requirements.txt
asn1crypto==0.23.0
attrs==19.3.0
Babel==2.5.1
bcrypt==3.1.4
boto3==1.4.7
botocore==1.7.43
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
Django==1.11.7
django-admin-sortable==2.2.3
django-autofixture==0.12.1
django-countries==5.0
django-debug-toolbar==1.8
django-storages==1.6.5
django-summernote==0.8.8.2
docutils==0.14
easypost==3.6.2
enum34==1.1.6
futures==3.1.1
idna==2.6
ipaddress==1.0.18
jmespath==0.9.3
olefile==0.44
paypalhttp==1.0.0
phonenumbers==8.12.4
Pillow==4.3.0
psycopg2==2.7.3.2
pycparser==2.18
pyOpenSSL==17.3.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
sorl-thumbnail==12.3
sqlparse==0.2.4
stripe==1.75.0
tenacity==6.2.0
urllib3==1.22
